+++ demandSide/src/optimalPrice/ToBeExpelled.java
@@ -1,8 +1,10 @@
 package optimalPrice;
 
+import java.math.BigDecimal;
+
 import firms.Firm;
 
 public class ToBeExpelled {
 	Firm f;
-	double priceToBeExpelled;
+	BigDecimal priceToBeExpelled;
 }
